refactor: extract AgentRunner.ToolExecution
Move-only extraction of the tool-call execution pipeline out of
Nous.AgentRunner into Nous.AgentRunner.ToolExecution (@moduledoc false,
internal): handle_tool_calls, sequential/parallel execution, pre-stage
decisions, approval + permission-policy enforcement, hook integration,
result recording, error formatting, and tool-call field helpers.
Task.Supervisor.async_stream_nolink under Nous.TaskSupervisor and the
ctx/run_ctx threading are preserved exactly. Telemetry events all remain
in the facade (none lived in moved code). No behavior change; public API
untouched.
